feat: extract files models to separate module
Extract Pydantic models from tools_files.py to files_models.py to enable faster imports and focused unit testing.

- Created src/inspect_agents/files_models.py with all result and parameter models
- Updated tools_files.py to import and re-export models from files_models
- Maintained full backward compatibility and import stability
- All 103 fs tests pass with no behavioral changes
