+++ prompts/FEATURE_PROMPT.txt
@@ -1,0 +1,52 @@
+# Source: prompts/features_by_phase/delivery-latest/phase-02.md (phase block)
+# Branch: feat/p02-gA-files-models-extraction-critical-path
+# Phase: 02
+# Group: A
+# Block: Files Models Extraction (Critical Path)
+# Features: feat-002-files-models-extraction
+
+## Group A: Files Models Extraction (Critical Path)
+
+Merged Features
+- feat-002-files-models-extraction — Extract Files result/param models
+
+Why merge
+- Single feature in this phase. Sets the base for later extractions.
+
+Dependencies & Gates (Block)
+- Hard deps: none
+- Soft deps: none
+- Phase gates: Phase 1 merged; re-exports in place.
+
+Implementation Guidance (02a‑style)
+- Context & Motivation: separate Pydantic models from I/O-heavy logic to speed imports and enable focused unit tests.
+- Steps
+  1) Create `src/inspect_agents/files_models.py` and move:
+     - Result models: `FileReadResult`, `FileWriteResult`, `FileEditResult`, `FileDeleteResult`, `FileTrashResult`, `FileListResult`, `FileMoveResult`, `FileStatResult`.
+     - Param models: `FilesParams`, `LsParams`, `ReadParams`, `WriteParams`, `EditParams`, `MkdirParams`, `MoveParams`, `StatParams`, `DeleteParams`, `TrashParams`.
+  2) In `src/inspect_agents/tools_files.py`: import-from and re-export the same names; keep `__all__` stable.
+  3) Grep identifiers to move: `class FileReadResult`, `class FilesParams`, `class ReadParams`, `LsParams`, etc.
+- Tests
+  - `tests/inspect_agents -k files_models`
+- Telemetry/Observability: none
+- Rollout & Guardrails: no behavior change; maintain import stability for downstream users.
+
+Scope Definition
+- In scope: `src/inspect_agents/tools_files.py`, `src/inspect_agents/files_models.py`, tests.
+- Out of scope: execution functions and sandbox/store logic.
+
+Success Criteria
+- Green tests: `files_models` selector.
+- Behavior assertions: both import paths resolve to the same classes (identity), no import cycles, import time not increased noticeably.
+
+Merge/Conflict Plan
+- Hot files: `src/inspect_agents/tools_files.py` (re-export edits).
+- Conflict mitigation: land quickly before Phase 3/4; small PR.
+
+Risk & Estimate (Block)
+- Aggregate risk: medium
+- Aggregate estimate: 2 points (M)
+
+Owner & Work Mode
+- Branch: `phase-02A-files-models`
+- Single PR; reviewers from files/tools area.
